[FEATURE] Generate main page
