+++ src/component/form/NameForm/index.jsx
@@ -5,7 +5,7 @@ function NameForm() {
   return (
     <>
       <p className="title">이름</p>
-      <input className="input" defaultValue={"이름을 입력하세요"}></input>
+      <input className="input" placeholder={"이름을 입력하세요"}></input>
     </>
   );
 }
