+++ src/pages/MainPage/index.css
@@ -1,0 +1,11 @@
+.title {
+  height: 20px;
+  color: black;
+  margin-bottom: 0;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
